+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 - [#133](https://github.com/influxdb/telegraf/pull/133): Add plugin.Gather error logging. Thanks @nickscript0!
 - [#136](https://github.com/influxdb/telegraf/issues/136): Add a -usage flag for printing usage of a single plugin.
+- [#137](https://github.com/influxdb/telegraf/issues/137): Memcached: fix when a value contains a space
 
 ### Bugfixes
 - [#128](https://github.com/influxdb/telegraf/issues/128): system_load measurement missing.